Try to fix the FreeBSD port to make it usable

+++ Pal/src/host/FreeBSD/db_events.c
@@ -78,7 +78,7 @@ int _DkEventSet (PAL_HANDLE event, int wakeup)
     return IS_ERR(ret) ? PAL_ERROR_TRYAGAIN : ret;
 }
 
-int _DkEventWaitTimeout (PAL_HANDLE event, int timeout)
+int _DkEventWaitTimeout (PAL_HANDLE event, uint64_t timeout)
 {
     int ret = 0;
     if (!event->event.isnotification || !atomic_read(&event->event.signaled)) {
